+++ ml4h/models/legacy_models.py
@@ -1330,87 +1330,6 @@ def make_paired_autoencoder_model(
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# ~~~~~~~ Training ~~~~~~~~~~~~~~~~
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-def train_model_from_generators(
-    model: Model,
-    generate_train: Iterable,
-    generate_valid: Iterable,
-    training_steps: int,
-    validation_steps: int,
-    batch_size: int,
-    epochs: int,
-    patience: int,
-    output_folder: str,
-    run_id: str,
-    inspect_model: bool,
-    inspect_show_labels: bool,
-    return_history: bool = False,
-    plot: bool = True,
-    save_last_model: bool = False,
-) -> Union[Model, Tuple[Model, History]]:
-    """Train a model from tensor generators for validation and training data.
-
-    Training data lives on disk, it will be loaded by generator functions.
-    Plots the metric history after training. Creates a directory to save weights, if necessary.
-    Measures runtime and plots architecture diagram if inspect_model is True.
-    :param model: The model to optimize
-    :param generate_train: Generator function that yields mini-batches of training data.
-    :param generate_valid: Generator function that yields mini-batches of validation data.
-    :param training_steps: Number of mini-batches in each so-called epoch
-    :param validation_steps: Number of validation mini-batches to examine after each epoch.
-    :param batch_size: Number of training examples in each mini-batch
-    :param epochs: Maximum number of epochs to run regardless of Early Stopping
-    :param patience: Number of epochs to wait before reducing learning rate.
-    :param output_folder: Directory where output file will be stored
-    :param run_id: User-chosen string identifying this run
-    :param inspect_model: If True, measure training and inference runtime of the model and generate architecture plot.
-    :param inspect_show_labels: If True, show labels on the architecture plot.
-    :param return_history: If true return history from training and don't plot the training history
-    :param plot: If true, plots the metrics for train and validation set at the end of each epoch
-    :param save_last_model: If true saves the model weights from last epoch otherwise saves model with best validation loss
-
-    :return: The optimized model.
-    """
-    model_file = os.path.join(output_folder, run_id, run_id + MODEL_EXT)
-    if not os.path.exists(os.path.dirname(model_file)):
-        os.makedirs(os.path.dirname(model_file))
-
-    if inspect_model:
-        image_p = os.path.join(output_folder, run_id, 'architecture_graph_' + run_id + IMAGE_EXT)
-        _inspect_model(model, generate_train, generate_valid, batch_size, training_steps, inspect_show_labels, image_p)
-
-    history = model.fit(
-        generate_train, steps_per_epoch=training_steps, epochs=epochs, verbose=1,
-        validation_steps=validation_steps, validation_data=generate_valid,
-        callbacks=_get_callbacks(patience, model_file, save_last_model),
-    )
-    generate_train.kill_workers()
-    generate_valid.kill_workers()
-
-    logging.info('Model weights saved at: %s' % model_file)
-    custom_dict = _get_custom_objects(generate_train.output_maps)
-    model = load_model(model_file, custom_objects=custom_dict, compile=False)
-    model.compile(optimizer='adam', loss=custom_dict['loss'])
-    if plot:
-        plot_metric_history(history, training_steps, run_id, os.path.dirname(model_file))
-    if return_history:
-        return model, history
-    return model
-
-
-def _get_callbacks(
-    patience: int, model_file: str, save_last_model: bool
-) -> List[Callback]:
-    callbacks = [
-        ModelCheckpoint(filepath=model_file, verbose=1, save_best_only=not save_last_model),
-        EarlyStopping(monitor='val_loss', patience=patience * 3, verbose=1),
-        ReduceLROnPlateau(monitor='val_loss', factor=0.5, patience=patience, verbose=1),
-    ]
-    return callbacks
-
-
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # ~~~~~~~ Predicting ~~~~~~~~~~~~~~
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 def embed_model_predict(model, tensor_maps_in, embed_layer, test_data, batch_size):
@@ -1496,136 +1415,6 @@ def _regularization_layer(dimension: int, regularization_type: str, rate: float)
     if regularization_type in DENSE_REGULARIZATION_CLASSES:
         return DENSE_REGULARIZATION_CLASSES[regularization_type](rate)
     return CONV_REGULARIZATION_CLASSES[regularization_type][dimension](rate)
-
-
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# ~~~~~~~ Inspections ~~~~~~~~~~~~~
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-def _inspect_model(
-    model: Model,
-    generate_train: Iterable,
-    generate_valid: Iterable,
-    batch_size: int,
-    training_steps: int,
-    inspect_show_labels: bool,
-    image_path: str,
-) -> Model:
-    """Collect statistics on model inference and training times.
-
-    Arguments:
-        model: the model to inspect
-        generate_train: training data generator function
-        generate_valid: Validation data generator function
-        batch_size: size of the mini-batches
-        training_steps: number of optimization steps to take
-        inspect_show_labels: if True, show layer labels on the architecture diagram
-        image_path: file path of the architecture diagram
-
-    Returns:
-        The slightly optimized keras model
-    """
-    if image_path:
-        _plot_dot_model_in_color(model_to_dot(model, show_shapes=inspect_show_labels, expand_nested=True), image_path, inspect_show_labels)
-    t0 = time.time()
-    _ = model.fit(generate_train, steps_per_epoch=training_steps, validation_steps=1, validation_data=generate_valid)
-    t1 = time.time()
-    n = batch_size * training_steps
-    train_speed = (t1 - t0) / n
-    logging.info(f'Spent:{(t1 - t0):0.2f} seconds training, Samples trained on:{n} Per sample training speed:{train_speed:0.3f} seconds.')
-
-    t0 = time.time()
-    inference_steps = max(1, training_steps//8)
-    _ = model.predict(generate_valid, steps=inference_steps, verbose=1)
-    t1 = time.time()
-    inference_speed = (t1 - t0) / (batch_size * inference_steps)
-
-    logging.info(f'Spent:{(t1 - t0):0.2f} seconds predicting, Samples inferred:{n} Per sample inference speed:{inference_speed:0.4f} seconds.')
-    return model
-
-
-def _plot_dot_model_in_color(dot, image_path, inspect_show_labels):
-    import pydot
-    legend = {}
-    for n in dot.get_nodes():
-        if n.get_label():
-            if 'Conv1' in n.get_label():
-                legend['Conv1'] = "cyan"
-                n.set_fillcolor("cyan")
-            elif 'Conv2' in n.get_label():
-                legend['Conv2'] = "deepskyblue1"
-                n.set_fillcolor("deepskyblue1")
-            elif 'Conv3' in n.get_label():
-                legend['Conv3'] = "deepskyblue3"
-                n.set_fillcolor("deepskyblue3")
-            elif 'UpSampling' in n.get_label():
-                legend['UpSampling'] = "darkslategray2"
-                n.set_fillcolor("darkslategray2")
-            elif 'Transpose' in n.get_label():
-                legend['Transpose'] = "deepskyblue2"
-                n.set_fillcolor("deepskyblue2")
-            elif 'BatchNormalization' in n.get_label():
-                legend['BatchNormalization'] = "goldenrod1"
-                n.set_fillcolor("goldenrod1")
-            elif 'output_' in n.get_label():
-                n.set_fillcolor("darkolivegreen2")
-                legend['Output'] = "darkolivegreen2"
-            elif 'softmax' in n.get_label():
-                n.set_fillcolor("chartreuse")
-                legend['softmax'] = "chartreuse"
-            elif 'MaxPooling' in n.get_label():
-                legend['MaxPooling'] = "aquamarine"
-                n.set_fillcolor("aquamarine")
-            elif 'Dense' in n.get_label():
-                legend['Dense'] = "gold"
-                n.set_fillcolor("gold")
-            elif 'Reshape' in n.get_label():
-                legend['Reshape'] = "coral"
-                n.set_fillcolor("coral")
-            elif 'Input' in n.get_label():
-                legend['Input'] = "darkolivegreen1"
-                n.set_fillcolor("darkolivegreen1")
-            elif 'Activation' in n.get_label():
-                legend['Activation'] = "yellow"
-                n.set_fillcolor("yellow")
-        n.set_style("filled")
-        if not inspect_show_labels:
-            n.set_label('\n')
-
-    for label in legend:
-        legend_node = pydot.Node('legend' + label, label=label, shape="box", fillcolor=legend[label])
-        dot.add_node(legend_node)
-
-    logging.info('Saving architecture diagram to:{}'.format(image_path))
-    dot.write_png(image_path)
-
-
-def saliency_map(input_tensor: np.ndarray, model: Model, output_layer_name: str, output_index: int) -> np.ndarray:
-    """Compute saliency maps of the given model (presumably already trained) on a batch of inputs with respect to the desired output layer and index.
-
-    For example, with a trinary classification layer called quality and classes good, medium, and bad output layer name would be "quality_output"
-    and output_index would be 0 to get gradients w.r.t. good, 1 to get gradients w.r.t. medium, and 2 for gradients w.r.t. bad.
-
-    :param input_tensor: A batch of input tensors
-    :param model: A trained model expecting those inputs
-    :param output_layer_name: The name of the output layer that the derivative will be taken with respect to
-    :param output_index: The index within the output layer that the derivative will be taken with respect to
-
-    :return: Array of the gradients same shape as input_tensor
-    """
-    get_gradients = _gradients_from_output(model, output_layer_name, output_index)
-    activation, gradients = get_gradients([input_tensor])
-    return gradients
-
-
-def _gradients_from_output(model, output_layer, output_index):
-    K.set_learning_phase(1)
-    input_tensor = model.input
-    x = model.get_layer(output_layer).output[:, output_index]
-    grads = K.gradients(x, input_tensor)[0]
-    grads /= (K.sqrt(K.mean(K.square(grads))) + 1e-6)  # normalization trick: we normalize the gradient
-    grads /= 80
-    iterate = K.function([input_tensor], [x, grads])
-    return iterate
 
 
 def _get_tensor_maps_for_characters(
